requirements.txt
# Cybersecurity ML Framework - Requirements

## Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
torch>=1.11.0
torchvision>=0.12.0

## Data Processing
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
scipy>=1.7.0

## Deep Learning
keras>=2.8.0
pytorch-lightning>=1.6.0
transformers>=4.20.0

## Federated Learning
flwr>=1.0.0
syft>=0.6.0

## Explainable AI
shap>=0.41.0
lime>=0.2.0
captum>=0.6.0

## Network Analysis
networkx>=2.6.0
scapy>=2.4.0
dpkt>=1.9.0

## Database and Storage
sqlite3
h5py>=3.6.0
joblib>=1.1.0

## Testing and Development
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

## Jupyter and Visualization
jupyter>=1.0.0
ipywidgets>=7.7.0
notebook>=6.4.0

## Utilities
tqdm>=4.64.0
click>=8.0.0
pyyaml>=6.0
python-dotenv>=0.19.0

## Performance Monitoring
psutil>=5.8.0
memory-profiler>=0.60.0

## Security and Privacy
cryptography>=3.4.0
pycryptodome>=3.15.0

## Optional GPU Support
cupy-cuda11x>=10.0.0  # For CUDA 11.x
cupy-cuda12x>=12.0.0  # For CUDA 12.x
